requirements.txt
streamlit>=1.30.0
psycopg[binary]>=3.1.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pandas>=2.0.0
